requirements.txt
pdfplumber
pathlib
io
anthropic
json
fastapi
uvicorn
python-dotenv